Introduce Aba::Batch class
Functionality was moved from `Aba` class. `Aba` class should contain
only a public API for any operations while dealing with ABA flat files.
Operations like preparing, creating or parsing ABA flat files should be
implemented in dedicated classes. This change also allows introducing a
class for parsing ABA flat files while keeping clear structure of gem
files.
