+++ README.md
@@ -10,7 +10,7 @@ Generates ABA (Australian Banking Association) file format output
 require 'aba'
 
 # Initialise ABA
-aba = Aba.new(
+aba = Aba.batch(
   bsb: "123-345", # Optional (Not required by NAB)
   financial_institution: "WPC",
   user_name: "John Doe",
@@ -44,18 +44,18 @@ There are a few ways to create a complete set of ABA data:
 
 ```ruby
 # Transactions added to the defined ABA object variable
-aba = Aba.new financial_institution: 'ANZ', user_name: 'Joe Blow', user_id: 123456, process_at: 200615
+aba = Aba.batch financial_institution: 'ANZ', user_name: 'Joe Blow', user_id: 123456, process_at: 200615
 aba.add_transaction bsb: '123-456', account_number: '000-123-456', amount: 50000
 aba.add_transaction bsb: '456-789', account_number: '123-456-789', amount: '-10000', transaction_code: 13
 
 # Transactions passed individually inside a block
-aba = Aba.new financial_institution: 'ANZ', user_name: 'Joe Blow', user_id: 123456, process_at: 200615 do |a|
+aba = Aba.batch financial_institution: 'ANZ', user_name: 'Joe Blow', user_id: 123456, process_at: 200615 do |a|
   a.add_transaction bsb: '123-456', account_number: '000-123-456', amount: 50000
   a.add_transaction bsb: '456-789', account_number: '123-456-789', amount: '-10000', transaction_code: 13
 end
 
-# Transactions as an array passed to the second param of Aba.new()
-aba = Aba.new(
+# Transactions as an array passed to the second param of Aba.batch
+aba = Aba.batch(
   { financial_institution: 'ANZ', user_name: 'Joe Blow', user_id: 123456, process_at: 200615 },
   [
     { bsb: '123-456', account_number: '000-123-456', amount: 50000 },
@@ -71,7 +71,7 @@ Validation errors can be caught in several ways:
 
 ```ruby
 # Create an ABA object with invalid character in the user_name
-aba = Aba.new(
+aba = Aba.batch(
   financial_institution: "ANZ",
   user_name: "Jøhn Doe",
   user_id: "123456",
